Report Module & Login

+++ src/components/Cliente/Cliente.jsx
@@ -1,32 +1,55 @@
 import React, { useEffect, useState } from "react";
 import axios from 'axios';
 import { Link } from "react-router-dom";
-
+import Swal from "sweetalert2";
 
 export default function Cliente() {
-    const [clientes, setClientes] = useState([]);
+  const [clientes, setClientes] = useState([]);
 
-    useEffect(() => {
-        console.log('Realizando solicitud GET...');
-        axios.get('http://localhost:6060/cliente')
-            .then(res => {
-                console.log('Solicitud GET exitosa:', res);
-                setClientes(res.data);
-            })
-            .catch(err => {
-                console.log('Error en la solicitud GET:', err);
-            });
-    }, []);
+  useEffect(() => {
+      console.log('Realizando solicitud GET...');
+      axios.get('http://localhost:6060/cliente')
+          .then(res => {
+              console.log('Solicitud GET exitosa:', res);
+              setClientes(res.data);
+          })
+          .catch(err => {
+              console.log('Error en la solicitud GET:', err);
+          });
+  }, []);
 
-    const handleDelete = async (id) => {
-        try {
+  const handleDelete = async (id) => {
+    try {
+        const result = await Swal.fire({
+            title: "Estás seguro?",
+            text: "No serás capaz de revertirlo!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            cancelButtonText: "Cancelar",
+            confirmButtonText: "Sí, eliminarlo!"
+        });
+
+        if (result.isConfirmed) {
             await axios.delete(`http://localhost:6060/cliente/${id}`);
-            window.location.reload();
-        } catch (err) {
-            console.log(err);
-            // Manejar el error de eliminación aquí
+            Swal.fire({
+                title: "Eliminado!",
+                text: "Eliminado exitosamente.",
+                icon: "success"
+            }).then(() => {
+                window.location.reload();
+            });
         }
+    } catch (err) {
+        console.log(err);
+        Swal.fire({
+            icon: "error",
+            title: "Oops...",
+            text: "Algo ha salido mal!",
+        });
     }
+}
     
 
     return (
